--- app/src/main/java/com/deloitte/project/benchmarking/data/api/ApiHelper.kt
@@ -1,6 +1,0 @@
-package com.deloitte.project.benchmarking.data.api
-
-class ApiHelper(private val apiService: ApiService) {
-
-    suspend fun getBlogs() = apiService.getBlogs()
-}
