--- app/src/main/java/com/deloitte/project/benchmarking/data/api/RetrofitBuilder.kt
@@ -1,18 +1,0 @@
-package com.deloitte.project.benchmarking.data.api
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object RetrofitBuilder {
-
-    private const val BASE_URL = "https://open-api.xyz/placeholder/"
-
-    private fun getRetrofit(): Retrofit {
-        return Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-    }
-
-    val apiService: ApiService = getRetrofit().create(ApiService::class.java)
-}
