--- app/src/main/java/com/deloitte/project/benchmarking/data/model/Blog.kt
@@ -1,9 +1,0 @@
-package com.deloitte.project.benchmarking.data.model
-
-data class Blog(
-    val body: String,
-    val category: String,
-    val image: String,
-    val pk: Int,
-    val title: String
-)
