--- app/src/main/java/com/deloitte/project/benchmarking/ui/main/viewmodel/MainViewModel.kt
@@ -1,20 +1,0 @@
-package com.deloitte.project.benchmarking.ui.main.viewmodel
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.liveData
-import com.deloitte.project.benchmarking.data.repository.MainRepository
-import com.deloitte.project.benchmarking.utils.Resource
-import kotlinx.coroutines.Dispatchers
-
-
-class MainViewModel(private val mainRepository: MainRepository) : ViewModel() {
-
-    fun getUsers() = liveData(Dispatchers.IO) {
-        emit(Resource.loading(data = null))
-        try {
-            emit(Resource.success(data = mainRepository.getBlogs()))
-        } catch (exception: Exception) {
-            emit(Resource.error(data = null, message = exception.message ?: "Error Occurred!"))
-        }
-    }
-}
