--- app/src/main/java/com/deloitte/project/benchmarking/utils/Resource.kt
@@ -1,17 +1,0 @@
-package com.deloitte.project.benchmarking.utils
-
-import com.deloitte.project.benchmarking.utils.Status.ERROR
-import com.deloitte.project.benchmarking.utils.Status.LOADING
-import com.deloitte.project.benchmarking.utils.Status.SUCCESS
-
-
-data class Resource<out T>(val status: Status, val data: T?, val message: String?) {
-    companion object {
-        fun <T> success(data: T): Resource<T> = Resource(status = SUCCESS, data = data, message = null)
-
-        fun <T> error(data: T?, message: String): Resource<T> =
-            Resource(status = ERROR, data = data, message = message)
-
-        fun <T> loading(data: T?): Resource<T> = Resource(status = LOADING, data = data, message = null)
-    }
-}
