--- app/src/main/java/com/deloitte/project/benchmarking/utils/Status.kt
@@ -1,7 +1,0 @@
-package com.deloitte.project.benchmarking.utils
-
-enum class Status {
-    SUCCESS,
-    ERROR,
-    LOADING
-}
