+++ app/src/main/java/com/sample/project/benchmarking/data/api/ApiHelper.kt
@@ -1,0 +1,6 @@
+package com.sample.project.benchmarking.data.api
+
+class ApiHelper(private val apiService: ApiService) {
+
+    suspend fun getBlogs() = apiService.getBlogs()
+}
