+++ app/src/main/java/com/sample/project/benchmarking/data/api/ApiService.kt
@@ -1,0 +1,11 @@
+package com.sample.project.benchmarking.data.api
+
+import com.sample.project.benchmarking.data.model.Blog
+import retrofit2.http.GET
+
+interface ApiService {
+
+    @GET("blogs")
+    suspend fun getBlogs(): List<Blog>
+
+}
