+++ app/src/main/java/com/sample/project/benchmarking/data/model/Blog.kt
@@ -1,0 +1,9 @@
+package com.sample.project.benchmarking.data.model
+
+data class Blog(
+    val body: String,
+    val category: String,
+    val image: String,
+    val pk: Int,
+    val title: String
+)
