+++ app/src/main/java/com/sample/project/benchmarking/data/repository/MainRepository.kt
@@ -1,0 +1,8 @@
+package com.sample.project.benchmarking.data.repository
+
+import com.sample.project.benchmarking.data.api.ApiHelper
+
+class MainRepository(private val apiHelper: ApiHelper) {
+
+    suspend fun getBlogs() = apiHelper.getBlogs()
+}
