+++ app/src/main/java/com/sample/project/benchmarking/ui/base/ViewModelFactory.kt
@@ -1,0 +1,19 @@
+package com.sample.project.benchmarking.ui.base
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import com.sample.project.benchmarking.data.api.ApiHelper
+import com.sample.project.benchmarking.data.repository.MainRepository
+import com.sample.project.benchmarking.ui.main.viewmodel.MainViewModel
+
+class ViewModelFactory(private val apiHelper: ApiHelper) : ViewModelProvider.Factory {
+
+    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(MainViewModel::class.java)) {
+            return MainViewModel(MainRepository(apiHelper)) as T
+        }
+        throw IllegalArgumentException("Unknown class name")
+    }
+
+}
+
