+++ app/src/main/java/com/sample/project/benchmarking/ui/main/view/MainActivity.kt
@@ -1,0 +1,83 @@
+package com.sample.project.benchmarking.ui.main.view
+
+import android.os.Bundle
+import android.view.View
+import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProviders
+import androidx.recyclerview.widget.DividerItemDecoration
+import androidx.recyclerview.widget.LinearLayoutManager
+import com.sample.project.benchmarking.R
+import com.sample.project.benchmarking.data.api.ApiHelper
+import com.sample.project.benchmarking.data.api.RetrofitBuilder
+import com.sample.project.benchmarking.data.model.Blog
+import com.sample.project.benchmarking.ui.base.ViewModelFactory
+import com.sample.project.benchmarking.ui.main.adapter.MainAdapter
+import com.sample.project.benchmarking.ui.main.viewmodel.MainViewModel
+import com.sample.project.benchmarking.utils.Status.*
+import kotlinx.android.synthetic.main.activity_main.*
+
+class MainActivity : AppCompatActivity() {
+
+    private lateinit var viewModel: MainViewModel
+    private lateinit var adapter: MainAdapter
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+        setupViewModel()
+        setupUI()
+        setupObservers()
+    }
+
+
+    private fun setupViewModel() {
+        viewModel = ViewModelProviders.of(
+            this,
+            ViewModelFactory(ApiHelper(RetrofitBuilder.apiService))
+        ).get(MainViewModel::class.java)
+    }
+
+    private fun setupUI() {
+        recyclerView.layoutManager = LinearLayoutManager(this)
+        adapter = MainAdapter(arrayListOf())
+        recyclerView.addItemDecoration(
+            DividerItemDecoration(
+                recyclerView.context,
+                (recyclerView.layoutManager as LinearLayoutManager).orientation
+            )
+        )
+        recyclerView.adapter = adapter
+    }
+
+    private fun setupObservers() {
+        viewModel.getUsers().observe(this, Observer {
+            it?.let { resource ->
+               when (resource.status) {
+                    SUCCESS -> {
+                        recyclerView.visibility = View.VISIBLE
+                        progressBar.visibility = View.GONE
+                        resource.data?.let { users -> retrieveList(users) }
+                    }
+                    ERROR -> {
+                        recyclerView.visibility = View.VISIBLE
+                        progressBar.visibility = View.GONE
+                        Toast.makeText(this, it.message, Toast.LENGTH_LONG).show()
+                    }
+                    LOADING -> {
+                        progressBar.visibility = View.VISIBLE
+                        recyclerView.visibility = View.GONE
+                    }
+                }
+            }
+        })
+    }
+
+    private fun retrieveList(blogs: List<Blog>) {
+        adapter.apply {
+            addUsers(blogs)
+            notifyDataSetChanged()
+        }
+    }
+}
