+++ app/src/main/java/com/sample/project/benchmarking/utils/Status.kt
@@ -1,0 +1,7 @@
+package com.sample.project.benchmarking.utils
+
+enum class Status {
+    SUCCESS,
+    ERROR,
+    LOADING
+}
